// app/interfaces/MainComponent/Enterprisebusiness.ts
export interface IEnterpriseBusiness {
    _id: string;
    id?: number;
    pagetitle?: string;
    pagedescription?: string;
    title?: string;
    image?: string;
    description?: string;
}

export interface IEnterpriseBusinessInput {
    id?: number;
    pagetitle?: string;
    pagedescription?: string;
    title?: string;
    image?: string;
    description?: string;
}